agent-runner/requirements.txt
litellm>=1.30.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
slowapi>=0.1.9
httpx>=0.26.0
pydantic>=2.5.0
